extracts Magic Link login to components/Authentication/EmailLogin.js
