+++ .env.bk
@@ -7,7 +7,7 @@ ORACLE_URL=http://localhost:8090
 OPENQ_API_URL=http://localhost:4000
 COIN_API_URL=http://localhost:8081
 BLOCK_EXPLORER_BASE_URL=https://mumbai.polygonscan.com
-PATS=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
+PATS=
 OPENQ_ID=5fbd39c6916b7efb63cc
 OPENQ_API_SSR_URL=http://openq-api:4000
 OPENQ_SUBGRAPH_SSR_HTTP_URL=http://graph_node:8000/subgraphs/name/openqdev/openq
